Align L&F of category and starts.
- Menu items are shown in Edit and Preview Mode
- Update is done without callback
