+++ app/src/main/res/menu/menu_note_list_view.xml
@@ -2,11 +2,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/menu_category"
-        android:orderInCategory="40"
-        android:title="@string/menu_change_category"
-        app:showAsAction="ifRoom" />
-    <item
         android:id="@+id/menu_favorite"
         android:orderInCategory="50"
         android:title="@string/menu_favorite"
@@ -17,6 +12,11 @@
         android:icon="@drawable/ic_eye_white_24dp"
         android:orderInCategory="80"
         android:title="@string/menu_preview"
+        app:showAsAction="ifRoom" />
+    <item
+        android:id="@+id/menu_category"
+        android:orderInCategory="100"
+        android:title="@string/menu_change_category"
         app:showAsAction="ifRoom" />
     <item
         android:id="@+id/menu_share"
